# docs/gui/openapi.json
{
  "components": {
    "schemas": {
...
      "FrameSnapshot": {
        "description": "Immutable state snapshot at discrete simulation tick t (starting at t=0).",
        "properties": {
          "active_agents": {
            "title": "Active Agents",
            "type": "integer"
          },
          "commitments": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Commitments",
            "type": "object"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictRecord"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "contracts": {
            "items": {
              "$ref": "#/components/schemas/ContractRecord"
            },
            "title": "Contracts",
            "type": "array"
          },
          "heat_grid": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Heat Grid",
            "type": "object"
          },
          "heat_provenance": {
            "default": "hindsight: executed future occupancy; not agent knowledge",
            "title": "Heat Provenance",
            "type": "string"
          },
          "is_unsolvable": {
            "default": false,
            "title": "Is Unsolvable",
            "type": "boolean"
          },
          "local_heat": {
            "items": {
              "$ref": "#/components/schemas/DecisionHeatRecord"
            },
            "title": "Local Heat",
            "type": "array"
          },
          "local_heat_omitted": {
            "default": 0,
            "minimum": 0.0,
            "title": "Local Heat Omitted",
            "type": "integer"
          },
          "local_observations": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/LocalObservation"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Observations"
          },
          "phase": {
            "default": "initial",
            "enum": [
              "initial",
              "post_move"
            ],
            "title": "Phase",
            "type": "string"
          },
          "planned_paths": {
            "additionalProperties": {
              "items": {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "type": "array"
              },
              "type": "array"
            },
            "title": "Planned Paths",
            "type": "object"
          },
          "positions": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Positions",
            "type": "object"
          },
          "solved_agents": {
            "title": "Solved Agents",
            "type": "integer"
          },
          "statuses": {
            "additionalProperties": {
              "enum": [
                "active",
                "reached",
                "parked",
                "disappeared",
                "deadlocked",
                "collided"
              ],
              "type": "string"
            },
            "title": "Statuses",
            "type": "object"
          },
          "targets": {
            "additionalProperties": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Targets",
            "type": "object"
          },
          "telemetry_available": {
            "default": false,
            "title": "Telemetry Available",
            "type": "boolean"
          },
          "tick": {
            "title": "Tick",
            "type": "integer"
          },
          "tokens": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Tokens",
            "type": "object"
          }
        },
        "required": [
          "tick",
          "active_agents",
          "solved_agents",
          "positions",
          "targets",
          "statuses",
          "tokens"
        ],
        "title": "FrameSnapshot",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStatusResponse": {
        "properties": {
          "attempt_id": {
            "title": "Attempt Id",
            "type": "string"
          },
          "attempt_index": {
            "default": 0,
            "title": "Attempt Index",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "definition_digest": {
            "title": "Definition Digest",
            "type": "string"
          },
          "effective_config": {
            "additionalProperties": true,
            "title": "Effective Config",
            "type": "object"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "parent_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Job Id"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SolverRunResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "solver_id": {
            "title": "Solver Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled",
              "timed_out",
              "interrupted"
            ],
            "title": "State",
            "type": "string"
          },
          "timeout_diagnostics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Diagnostics"
          },
          "timeout_scope": {
            "anyOf": [
              {
                "enum": [
                  "process",
                  "negotiation"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Scope"
          },
          "timeout_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Session Id"
          },
          "trial_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Id"
          },
          "worker_pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Pid"
          }
        },
        "required": [
          "job_id",
          "run_id",
          "attempt_id",
          "definition_digest",
          "solver_id",
          "state",
          "created_at",
          "effective_config"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "JobSubmissionRequest": {
        "additionalProperties": false,
        "properties": {
          "broadcast_horizon": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broadcast Horizon"
          },
          "commitment_type": {
            "default": "SC",
            "enum": [
              "SC",
              "DC",
              "ZC"
            ],
            "title": "Commitment Type",
            "type": "string"
          },
          "fov_size": {
            "default": 3,
            "maximum": 15.0,
            "minimum": 3.0,
            "title": "Fov Size",
            "type": "integer"
          },
          "goals": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "maxProperties": 100,
            "title": "Goals",
            "type": "object"
          },
          "grid_height": {
            "default": 8,
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Grid Height",
            "type": "integer"
          },
          "grid_width": {
            "default": 8,
            "maximum": 64.0,
            "minimum": 1.0,
            "title": "Grid Width",
            "type": "integer"
          },
          "heat_recording_limit": {
            "default": 250000,
            "maximum": 2000000.0,
            "minimum": 0.0,
            "title": "Heat Recording Limit",
            "type": "integer"
          },
          "initial_tokens": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Initial Tokens",
            "type": "integer"
          },
          "max_astar_expansions": {
            "default": 1500,
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Max Astar Expansions",
            "type": "integer"
          },
          "max_steps": {
            "default": 100,
            "description": "Explicit simulation step guard; interactive form caps at 500, research batches permit 10000",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "name": {
            "default": "Custom scenario",
            "maxLength": 160,
            "title": "Name",
            "type": "string"
          },
          "negotiation_deadline_sec": {
            "default": 60.0,
            "description": "Wall seconds for each bilateral negotiation; resets per session, not per offer",
            "maximum": 600.0,
            "minimum": 0.01,
            "title": "Negotiation Deadline Sec",
            "type": "number"
          },
          "negotiation_horizon": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negotiation Horizon"
          },
          "negotiation_protocol": {
            "default": "taop-v2",
            "enum": [
              "taop-v2",
              "taop-v1",
              "legacy-alternating-v1"
            ],
            "title": "Negotiation Protocol",
            "type": "string"
          },
          "negotiation_round_limit": {
            "default": 30,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Negotiation Round Limit",
            "type": "integer"
          },
          "obstacles": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "maxItems": 4096,
            "title": "Obstacles",
            "type": "array"
          },
          "random_seed": {
            "default": 42,
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Random Seed",
            "type": "integer"
          },
          "recording_level": {
            "default": "full-trace",
            "enum": [
              "metrics-only",
              "events",
              "full-trace"
            ],
            "title": "Recording Level",
            "type": "string"
          },
          "scenario_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "setting": {
            "default": "SETTING_1",
            "enum": [
              "SETTING_1",
              "SETTING_2",
              "SETTING_3",
              "SETTING_4"
            ],
            "title": "Setting",
            "type": "string"
          },
          "solver_id": {
            "default": "Decentralized-HeatMap",
            "title": "Solver Id",
            "type": "string"
          },
          "starts": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "maxProperties": 100,
            "title": "Starts",
            "type": "object"
          },
          "suboptimality": {
            "default": 1.1,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Suboptimality",
            "type": "number"
          },
          "timeout_sec": {
            "anyOf": [
              {
                "maximum": 600.0,
                "minimum": 0.1,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 5.0,
            "description": "Optional whole-process safety budget; null removes this cap for decentralized jobs only",
            "title": "Timeout Sec"
          },
          "verification_pass_limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Verification Pass Limit",
            "type": "integer"
          }
        },
        "title": "JobSubmissionRequest",
        "type": "object"
      },
...
      "Point2D": {
        "properties": {
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "Point2D",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": true,
        "properties": {
          "frames": {
            "items": {
              "$ref": "#/components/schemas/FrameSnapshot"
            },
            "title": "Frames",
            "type": "array"
          },
          "metadata": {
            "$ref": "#/components/schemas/RunMetadata"
          },
          "result": {
            "$ref": "#/components/schemas/SolverRunResult"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "metadata",
          "result",
          "frames"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "SolverRunResult": {
        "description": "Complete execution record and replay payload for a single solver branch.",
        "properties": {
          "frames": {
            "items": {
              "$ref": "#/components/schemas/FrameSnapshot"
            },
            "title": "Frames",
            "type": "array"
          },
          "information_sharing_rate": {
            "default": 0.0,
            "title": "Information Sharing Rate",
            "type": "number"
          },
          "instance_hash": {
            "title": "Instance Hash",
            "type": "string"
          },
          "is_centralized": {
            "title": "Is Centralized",
            "type": "boolean"
          },
          "makespan": {
            "title": "Makespan",
            "type": "integer"
          },
          "measured_metrics": {
            "additionalProperties": true,
            "title": "Measured Metrics",
            "type": "object"
          },
          "metric_availability": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Metric Availability",
            "type": "object"
          },
          "negotiation_count": {
            "default": 0,
            "title": "Negotiation Count",
            "type": "integer"
          },
          "norm_path_diff": {
            "default": 0.0,
            "title": "Norm Path Diff",
            "type": "number"
          },
          "paths": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Point2D"
              },
              "type": "array"
            },
            "title": "Paths",
            "type": "object"
          },
          "reported_makespan": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Makespan"
          },
          "reported_sum_of_costs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Sum Of Costs"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "runtime_ms": {
            "title": "Runtime Ms",
            "type": "number"
          },
          "solved_count": {
            "title": "Solved Count",
            "type": "integer"
          },
          "solver": {
            "title": "Solver",
            "type": "string"
          },
          "solver_key": {
            "title": "Solver Key",
            "type": "string"
          },
          "solver_outcome": {
            "default": "not_solved",
            "enum": [
              "solved",
              "not_solved"
            ],
            "title": "Solver Outcome",
            "type": "string"
          },
          "status": {
            "enum": [
              "solved",
              "failed",
              "truncated",
              "invalid"
            ],
            "title": "Status",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "successful_negotiations": {
            "default": 0,
            "title": "Successful Negotiations",
            "type": "integer"
          },
          "sum_of_costs": {
            "default": 0,
            "title": "Sum Of Costs",
            "type": "integer"
          },
          "telemetry_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Telemetry Events",
            "type": "array"
          },
          "total_agents": {
            "title": "Total Agents",
            "type": "integer"
          },
          "validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ValidationReceipt"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "solver",
          "solver_key",
          "is_centralized",
          "status",
          "success",
          "solved_count",
          "total_agents",
          "makespan",
          "runtime_ms",
          "instance_hash",
          "run_id",
          "paths",
          "frames"
        ],
        "title": "SolverRunResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationReceipt": {
        "properties": {
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "first_violation_tick": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Violation Tick"
          },
          "is_valid": {
            "title": "Is Valid",
            "type": "boolean"
          },
          "prefix_valid": {
            "title": "Prefix Valid",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "valid_solution",
              "valid_prefix",
              "invalid",
              "not_checked"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "const": "trajectory-v2",
            "default": "trajectory-v2",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "is_valid",
          "prefix_valid"
        ],
        "title": "ValidationReceipt",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Decentralized MAPF simulation and centralized solver comparisons, independent trajectory validation, replay and experiment analysis.",
    "title": "DEC-MAPF - Research Workspace",
    "version": "0.1.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Jobs Api V1 Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Jobs"
      },
      "post": {
        "operationId": "submit_api_v1_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit"
      }
    },
...
    "/api/v1/runs/import": {
      "post": {
        "operationId": "import_run_api_v1_runs_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Bundle",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Import Run Api V1 Runs Import Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Run"
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "delete": {
        "operationId": "delete_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Api V1 Runs  Run Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete"
      },
      "get": {
        "operationId": "run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_frames",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Frames",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run"
      }
    },
...
  }
}